// src/store/mutation-types.ts
export enum MutationTypes {
  SET_DARK_MODE = "SET_DARK_MODE",
  SET_LUNA_PICTURES = "SET_LUNA_PICTURES",
  SET_LUNA_PICTURES_LOADING = "SET_LUNA_PICTURES_LOADING",
  SELECT_LUNA_PICTURE = "SELECT_LUNA_PICTURE",
  // Manga
  SET_MANGA_LOADING = "SET_MANGA_LOADING",
  SET_MANGA_DATA = "SET_MANGA_DATA",
  SET_MANGA_SELECTED = "SET_MANGA_SELECTED",
  SET_CHAPTER_SELECTED = "SET_CHAPTER_SELECTED",
  SET_PAGE_SELECTED = "SET_PAGE_SELECTED",
  SET_PAGE_SELECTED_INDEX = "SET_PAGE_SELECTED_INDEX"
}